+++ packages/adapter-pg/src/index.js
@@ -1,0 +1,28 @@
+const { AdapterError, ADAPTER_ERROR_CATALOG, throwAdapterError } = require("./errors");
+const { prepareMapping, ensurePreparedMapping } = require("./mapping");
+const {
+  OPERATOR_TEMPLATES,
+  getTableSql,
+  compileWhere,
+  compileOrderBy,
+  compileLimitOffset,
+  compileSecurityPredicate,
+  compileSelect
+} = require("./compile");
+const { assembleSelectSql } = require("./assemble");
+
+module.exports = {
+  AdapterError,
+  ADAPTER_ERROR_CATALOG,
+  throwAdapterError,
+  OPERATOR_TEMPLATES,
+  prepareMapping,
+  ensurePreparedMapping,
+  getTableSql,
+  compileWhere,
+  compileOrderBy,
+  compileLimitOffset,
+  compileSecurityPredicate,
+  compileSelect,
+  assembleSelectSql
+};